Mail/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pypdf
requests
google-generativeai
python-dotenv
